Fixed tons conversion warnings

+++ benchmark/hd_erase_d_benchmark.cc
@@ -42,7 +42,7 @@ class IndexBenchmark {
     void Remove(benchmark::State& state, PhTreeD<DIM, int>& tree);
 
     const TestGenerator data_type_;
-    const int num_entities_;
+    const size_t num_entities_;
 
     std::default_random_engine random_engine_;
     std::uniform_real_distribution<> cube_distribution_;
@@ -97,7 +97,7 @@ void IndexBenchmark<DIM>::Insert(benchmark::State&, PhTreeD<DIM, int>& tree) {
 
 template <dimension_t DIM>
 void IndexBenchmark<DIM>::Remove(benchmark::State& state, PhTreeD<DIM, int>& tree) {
-    int n = 0;
+    size_t n = 0;
     for (int i = 0; i < num_entities_; ++i) {
         n += tree.erase(points_[i]);
     }
