+++ benchmark/query_d_benchmark.cc
@@ -64,7 +64,7 @@ class IndexBenchmark {
     const double avg_query_result_size_;
 
     constexpr int query_endge_length() {
-        return GLOBAL_MAX * pow(avg_query_result_size_ / (double)num_entities_, 1. / (double)DIM);
+        return (int)(GLOBAL_MAX * pow(avg_query_result_size_ / (double)num_entities_, 1. / (double)DIM));
     };
 
     TreeType<DIM> tree_;
@@ -145,7 +145,7 @@ size_t Count_MMFE(TreeType<DIM>& tree, BoxType<DIM>& query_box) {
 
 template <dimension_t DIM, QueryType QUERY_TYPE>
 void IndexBenchmark<DIM, QUERY_TYPE>::QueryWorld(benchmark::State& state, BoxType<DIM>& query_box) {
-    int n = 0;
+    size_t n = 0;
     switch (QUERY_TYPE) {
     case MIN_MAX_ITER:
         n = Count_MMI(tree_, query_box);
